Update readme.md
...
God, it's been a while, hasn't it?
This was my first experience with Deno, and it *shows*. It's strange to think that my journey with it all started here, with such a crappy little thing.
I should probably revise this thing someday. Not now, but maybe later. I hear `std/walk` is a pretty suitable tool, for what I was trying to achieve back then...